Clean up fake Haml filters
