+++ test/apps/rails5.2/app/views/users/one.html.haml
@@ -6,3 +6,14 @@
 %p#name
   %b Name:
   = @user.name
+
+:coffeescript
+  $('h1').click ->
+  alert('You clicked on the headline')
+
+:markdown
+  # Jello!
+
+:sass
+  #div
+    font-weight: bold
